requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.1.3
billiard==4.2.1
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
cryptography==42.0.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
locust==2.31.8
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
orjson==3.10.3
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-sqlalchemy==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==26.2.0
redis==5.0.8
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
